[llvm-ar][test] Add tests failing on Darwin
These tests that failed on Darwin but passed on other machines due to the default archive format differing
on a Darwin machine, and what looks to be bugs in the output of this format.
I can not investigate these issue further so the tests are considered expected failures on Darwin.

Differential Revision: https://reviews.llvm.org/D64802

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@366334 91177308-0d34-0410-b5e6-96231b3b80d8