[NFC][InstCombine] Some tests for sub-of-negatible pattern
As we have previously estabilished, `sub` is an outcast,
and should be considered non-canonical iff it can be converted to `add`.

It can be converted to `add` if it's second operand can be negated.
So far we mostly only do that for constants and negation itself,
but we should be more through.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@373597 91177308-0d34-0410-b5e6-96231b3b80d8